Print AI Calculation Times & Phasing Out PrevBoards
A major reason for the inefficiency was due to each board containing each of its previous states. Every time a board was deepcopied, each additional previous state increased the copy time, which really adds up when the AI is looking at over 10,000 different states. By having the board store only one previous state, AI calculation times for a lookahead of 3 dropped from 87 seconds to 57 seconds. If we could figure out how to drop prevBoards entirely and still determine en_passent, then that'll further improve efficiency. Of course, this would remove the ability to undo moves, but I think that's a secondary concern at this point. If we do get around to it, we can always store the previous boards in a global variable.